Makes the handlers of mouseMove and touchMove events not passive.
Since Chrome 56 touch event handlers are `passive` by default.
https://www.chromestatus.com/features/5093566007214080

This makes the events uncancelable. In order to prevent a scrolling
element that contains the draggables from scrolling when interacting
with the draggables the touch events need to be canceled.

Binding the eventListeners with the { passive: false } option makes
that possible, again.

Btw i tried to set the `touch-action` on the scrolling container
in the touchStart handler but wasn't able to get that to execute
before the scrolling starts.